Replace NAT mode with local SOCKS5 mode. #1441
