+++ .gitmodules
@@ -9,14 +9,6 @@
 	path = mobile/src/main/jni/libancillary
 	url = https://github.com/shadowsocks/libancillary.git
 	branch = shadowsocks-android
-[submodule "mobile/src/main/jni/libevent"]
-	path = mobile/src/main/jni/libevent
-	url = https://github.com/shadowsocks/libevent.git
-	branch = shadowsocks-android
-[submodule "mobile/src/main/jni/redsocks"]
-	path = mobile/src/main/jni/redsocks
-	url = https://github.com/shadowsocks/redsocks.git
-	branch = shadowsocks-android
 [submodule "mobile/src/main/jni/mbedtls"]
 	path = mobile/src/main/jni/mbedtls
 	url = https://github.com/ARMmbed/mbedtls
