+++ mobile/src/main/jni/Android.mk
@@ -79,32 +79,6 @@ LOCAL_SRC_FILES := $(addprefix libsodium/src/libsodium/,$(SODIUM_SOURCE))
 include $(BUILD_STATIC_LIBRARY)
 
 ########################################################
-## libevent
-########################################################
-
-include $(CLEAR_VARS)
-
-LIBEVENT_SOURCES := \
-	buffer.c \
-	bufferevent.c bufferevent_filter.c \
-	bufferevent_pair.c bufferevent_ratelim.c \
-	bufferevent_sock.c epoll.c \
-	epoll_sub.c evdns.c event.c \
-    event_tagging.c evmap.c \
-	evrpc.c evthread.c \
-	evthread_pthread.c evutil.c \
-	evutil_rand.c http.c \
-	listener.c log.c poll.c \
-	select.c signal.c strlcpy.c
-
-LOCAL_MODULE := event
-LOCAL_SRC_FILES := $(addprefix libevent/, $(LIBEVENT_SOURCES))
-LOCAL_CFLAGS := -O2 -D_EVENT_HAVE_ARC4RANDOM -I$(LOCAL_PATH)/libevent \
-	-I$(LOCAL_PATH)/libevent/include \
-
-include $(BUILD_STATIC_LIBRARY)
-
-########################################################
 ## libancillary
 ########################################################
 
@@ -199,28 +173,6 @@ LOCAL_SRC_FILES := \
 	libev/event.c
 
 include $(BUILD_STATIC_LIBRARY)
-
-########################################################
-## redsocks
-########################################################
-
-include $(CLEAR_VARS)
-
-REDSOCKS_SOURCES := base.c http-connect.c \
-	log.c md5.c socks5.c \
-	base64.c http-auth.c http-relay.c main.c \
-	parser.c redsocks.c socks4.c utils.c
-
-LOCAL_STATIC_LIBRARIES := libevent
-
-LOCAL_MODULE := redsocks
-LOCAL_SRC_FILES := $(addprefix redsocks/, $(REDSOCKS_SOURCES)) 
-LOCAL_CFLAGS := -O2 -std=gnu99 -DUSE_IPTABLES \
-	-I$(LOCAL_PATH)/redsocks \
-	-I$(LOCAL_PATH)/libevent/include \
-	-I$(LOCAL_PATH)/libevent
-
-include $(BUILD_SHARED_EXECUTABLE)
 
 ########################################################
 ## shadowsocks-libev local
