+++ mobile/src/main/res/xml/pref_global.xml
@@ -8,6 +8,6 @@
                       android:summary="@string/tcp_fastopen_summary"
                       android:title="TCP Fast Open"/>
     <SwitchPreference android:key="isNAT"
-                      android:title="@string/nat"
-                      android:summary="@string/nat_summary"/>
+                      android:title="@string/local"
+                      android:summary="@string/local_summary"/>
 </PreferenceScreen>
