+++ mobile/src/main/scala/com/github/shadowsocks/utils/Utils.scala
@@ -33,7 +33,7 @@ import android.view.View.MeasureSpec
 import android.view.{Gravity, View, Window}
 import android.widget.Toast
 import com.github.shadowsocks.ShadowsocksApplication.app
-import com.github.shadowsocks.{BuildConfig, ShadowsocksNatService, ShadowsocksVpnService}
+import com.github.shadowsocks.{BuildConfig, ShadowsocksLocalService, ShadowsocksVpnService}
 import org.xbill.DNS._
 
 import scala.collection.JavaConversions._
@@ -138,7 +138,7 @@ object Utils {
 
   def startSsService(context: Context) {
     val intent =
-      new Intent(context, if (app.dataStore.isNAT) classOf[ShadowsocksNatService] else classOf[ShadowsocksVpnService])
+      new Intent(context, if (app.dataStore.isNAT) classOf[ShadowsocksLocalService] else classOf[ShadowsocksVpnService])
     if (Build.VERSION.SDK_INT >= 26) context.startForegroundService(intent) else context.startService(intent)
   }
   def reloadSsService(context: Context): Unit = context.sendBroadcast(new Intent(Action.RELOAD))
